convert "floatinghelp" to a higher level component. works in its original place but i dont see why its not working in the new location yet
