+++ src/components/InfoIconWithTooltip/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./InfoIconWithTooltip";
